+++ src/app/products/services/products.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
+
 import { ProductModel } from '../models/product.model';
-import { Category } from '../../common/enums/category';
 import { ProductsServicesModule } from '../products.services.module';
 
 @Injectable({
@@ -8,47 +9,49 @@ import { ProductsServicesModule } from '../products.services.module';
 })
 export class ProductsService {
 
-  private books: ProductModel[] = [
-    {
-      productId: 1,
-      name: 'Game of Thrones',
-      author: 'G. Martin',
-      category: Category.Fantasy,
-      price: 10,
-      isAvailable: true
-    },
-    {
-      productId: 2,
-      name: 'Dance with the Winds',
-      author: 'G. Martin',
-      category: Category.Fantasy,
-      price: 30,
-      isAvailable: true
-    }
-  ];
+  private booksUrl = 'http://localhost:3000/books';
+
+  constructor(private http: HttpClient) {}
 
   getBooks(): Promise<ProductModel[]> {
-    return Promise.resolve(this.books);
+    return this.http
+      .get(this.booksUrl)
+      .toPromise()
+      .then(response => <ProductModel[]>response);
   }
 
   getBookById(id: number): Promise<ProductModel> {
-    return this.getBooks().then((books) => {
-      return books.find(book => book.productId === id);
-    });
+    const url = `${this.booksUrl}/${id}`;
+    return this.http
+      .get(url)
+      .toPromise()
+      .then(response => response as ProductModel);
   }
 
-  addBook(book: ProductModel): Promise<void> {
-    return this.getBooks().then((books) => {
-      book.productId = books[books.length - 1].productId++;
-      books.push(book);
-    });
+  addBook(book: ProductModel): Promise<ProductModel> {
+    const url = this.booksUrl;
+    const body = JSON.stringify(book);
+    const options = {
+      headers: new HttpHeaders({ 'Content-Type': 'application/json' })
+    };
+
+    return this.http
+      .post(url, body, options)
+      .toPromise()
+      .then(response => response as ProductModel);
   }
 
-  updateBook(update: ProductModel): Promise<void> {
-    const i = this.books.findIndex(book => book.productId === update.productId);
-    if (i > -1) {
-      this.books.splice(i, 1, update);
-    }
-    return Promise.resolve();
+  updateBook(update: ProductModel): Promise<ProductModel> {
+
+    const url = `${this.booksUrl}/${update.id}`;
+    const body = JSON.stringify(update);
+    const options = {
+      headers: new HttpHeaders({ 'Content-Type': 'application/json' })
+    };
+
+    return this.http
+      .put(url, body, options)
+      .toPromise()
+      .then(response => response as ProductModel);
   }
 }